+++ docs/changelog.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Added
+
+* Units definition file at `data/units.txt`
+
+### Changed
+
+* Get unit registry from application and add Joseki required units by loading 
+  a units definition file.
+ 
 ## [2.4.0] - 2023-08-07
 
 ### Added